Prevent Nelio AB Testing recognising staging environments
The `nab_staging_urls` returns an array of strings that Nelio AB will
look for in the URL of the site where it's running. If it finds a match,
it treats that site as a staging environment, and disables some of its
functionality.

By returning an empty array, we stop that happening.